requirements.txt
python-dotenv
langgraph
langchain
langchain-google-genai
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
